move blueprint into apps folder
