+++ busy_beaver/apps/github_integration/api/oauth.py
@@ -3,9 +3,7 @@ import logging
 from flask import jsonify, request
 from flask.views import MethodView
 
-from busy_beaver.apps.oauth_integrations.github.workflow import (
-    process_github_oauth_callback,
-)
+from busy_beaver.apps.github_integration.workflow import process_github_oauth_callback
 
 logger = logging.getLogger(__name__)
